misc: Remove dev mode and config choices from the Menu
For underage children, they don't need to have access to the dev tools.
Nor do they need to have direct access to modify the config.

Yes, it is protection by obscurity. It should not be hard if someone
really wants to get around this with Ctrl+Shift+I or accessing the
config file directly